2012-08-28 07:09 UTC+0700 Andi Jahja <xharbour/AT/telkom/net/id>   * contrib/jpeg/jcmainct.c   * contrib/jpeg/jdmainct.c     ! variable renamed: main -> jmain
  * contrib/tiff/tif_config.h
  * contrib/tiff/tif_dirread.c
  * contrib/tiff/tif_dirwrite.c
  * contrib/tiff/tif_predict.c
  * contrib/tiff/tif_read.c
  * contrib/pdflite/tiff/tif_unix.c
    ! pacify GCC warnings

+++ xharbour/contrib/tiff/tif_predict.c
@@ -611,9 +611,9 @@ PredictorEncodeTile(TIFF* tif, uint8* bp0, tmsize_t cc0, uint16 s)
         working_copy = (uint8*) _TIFFmalloc(cc0);
         if( working_copy == NULL )
         {
-            TIFFErrorExt(tif->tif_clientdata, module, 
+            TIFFErrorExt(tif->tif_clientdata, module,
                          "Out of memory allocating " TIFF_SSIZE_FORMAT " byte temp buffer.",
-                         cc0 );
+                         (long int) cc0 );
             return 0;
         }
         memcpy( working_copy, bp0, cc0 );
